File copy refactoring
