+++ src/main/java/com/example/apgw/helper/FileStorageHelper.java
@@ -1,6 +1,5 @@
 package com.example.apgw.helper;
 
-import com.example.apgw.model.Assignment;
 import com.example.apgw.model.Submission;
 import org.apache.commons.io.FileUtils;
 import org.apache.commons.io.FilenameUtils;
@@ -8,14 +7,9 @@ import org.springframework.web.multipart.MultipartFile;
 
 import java.io.File;
 import java.io.IOException;
-import java.net.URISyntaxException;
 import java.nio.file.FileSystemException;
-import java.nio.file.Files;
 import java.nio.file.Path;
-import java.nio.file.Paths;
-import java.util.*;
-
-import static java.nio.file.StandardCopyOption.REPLACE_EXISTING;
+import java.util.Objects;
 
 public class FileStorageHelper {
 
@@ -66,60 +60,7 @@ public class FileStorageHelper {
         questionFile.transferTo(questionDest);
     }
 
-    /**
-     * Copy files to temp directory.
-     *
-     * @param submission submission whose files are to be copied.
-     * @param assignment assignment whose files are to be copied.
-     * @param path       path of tempdir.
-     * @throws IOException        If I/O fails.
-     * @throws URISyntaxException If path syntax is incorrect.
-     */
-    public void copyFiles(Submission submission, Assignment assignment, Path path)
-            throws IOException, URISyntaxException {
-        //create dir if does not exist
-        if (!path.toFile().exists()) {
-            boolean mkdir = path.toFile().mkdirs();
-            if (!mkdir) {
-                throw new FileSystemException("Cannot create dir");
-            }
-        }
-        //copy files to temp
-        Path submissionDirectory = Paths.get(basedir + "/apgw/submission/" + submission.getId());
-        Optional<Path> submissionPath = Files.list(submissionDirectory).findFirst();
-        Path inputPath = Paths.get(basedir + "/apgw/assi/" + assignment.getId() + "/input");
-        Path outputPath = Paths.get(basedir + "/apgw/assi/" + assignment.getId() + "/output");
-        if (submissionPath.isPresent()) {
-            Files.copy(submissionPath.get(),
-                    path.resolve(submissionPath.get().getFileName()),
-                    REPLACE_EXISTING);
-        }
-        Files.copy(inputPath, path.resolve("input"), REPLACE_EXISTING);
-        Files.copy(outputPath, path.resolve("output"), REPLACE_EXISTING);
 
-        //check type and copy appropriate script
-        String type = getCodeType(submission);
-        String scriptName;
-        switch (type) {
-            case "c":
-                scriptName = "c-script.sh";
-                break;
-            case "cpp":
-                scriptName = "cpp-script.sh";
-                break;
-            default:
-                scriptName = "";
-                break;
-        }
-
-        Files.copy(Paths.get(FileStorageHelper.class.getResource("/" + scriptName).toURI()),
-                path.resolve(scriptName),
-                REPLACE_EXISTING);
-        Boolean isExecutable = path.resolve(scriptName).toFile().setExecutable(true);
-        if (!isExecutable) {
-            throw new FileSystemException("cannot change permission");
-        }
-    }
 
     /**
      * Stores submission files to disk.
@@ -143,29 +84,6 @@ public class FileStorageHelper {
         String submissionPath = path + file.getOriginalFilename();
         File dest = new File(submissionPath);
         file.transferTo(dest);
-    }
-
-    /**
-     * Get the source code type of a submission.
-     *
-     * @param submission Submission to be checked.
-     * @return type, valid options are "c", "cpp" and "invalid".
-     */
-    public String getCodeType(Submission submission) {
-        File codeDir = new File(basedir + "/apgw/submission/" + submission.getId());
-        File codeFile = Objects.requireNonNull(codeDir.listFiles())[0];
-        String type = FilenameUtils.getExtension(codeFile.getName());
-        if (type.equals("c")) {
-            return "c";
-        }
-
-        //c++ types
-        Set<String> cppSuffixes =
-                new HashSet<>(Arrays.asList("cc", "cp", "cxx", "cpp", "CPP", "c++", "C"));
-        if (cppSuffixes.contains(type)) {
-            return "cpp";
-        }
-        return "invalid";
     }
 
     /**
